app/requirements.txt
jupyter
cassandra-driver
pandas
python-dotenv